+++ tests/test_order.py
@@ -16,7 +16,6 @@ ORDER_PLACEMENT_URL = f"{constants.API_HOST}/v2/orders?workspaceId={constants.WO
 def workspace():
     with mock.patch("up42.base.workspace") as workspace_mock:
         workspace_mock.auth.session = requests.session()
-        workspace_mock.id = constants.WORKSPACE_ID
         yield