+++ up42/order.py
@@ -107,7 +107,6 @@ class Order:
     """
 
     session = base.Session()
-    workspace_id = base.WorkspaceId()
 
     def __init__(
         self,